+++ CHANGELOG.md
@@ -2,7 +2,9 @@ Important changes to data models, configuration, and migrations between each
 AppEngine version, listed here to ease deployment and troubleshooting.
 
 ## Next Release (replace with git tag when deployed)
- * Bump runtimeVersion to `2026.03.07`.
+ * Bump runtimeVersion to `2026.03.10`.
+ * Upgraded stable Dart analysis SDK to `3.11.2`
+ * Upgraded stable Flutter analysis SDK to `3.41.4`.
  * Note: Updated the scaling settings of the `search` service instances.
 
 ## `20260306t095100-all`